Huge refactor and cleanout
