+++ include/tasks/queues/QueueFactory.h
@@ -1,7 +1,6 @@
 
 #pragma once
 
-#include <ControllerClass.h>
 #include <tasks/queues/queues.h>
 
 #include <QueueManager.h>
@@ -21,10 +20,6 @@ Queue1::Manager<T> *createQueue(const char *name, TickType_t ticks = TICKS_5ms)
   {
     return new Queue1::Manager<T>(xSimplMessageQueue, TICKS_5ms, name);
   }
-  // if (std::is_same<T, DisplayEvent>::value)
-  // {
-  //   return new Queue1::Manager<T>(xDisplayQueueHandle, TICKS_5ms, name);
-  // }
   Serial.printf("ERROR: (Manager::create) a queue has not been created for this type (%s)\n", name);
   return nullptr;
 }
