+++ include/tasks/root.h
@@ -1,20 +1,23 @@
 
 #include <tasks/core_0/ControllerCommsTask.h>
+#include <tasks/core_0/HeadlightTask.h>
+#include <tasks/core_0/I2CPortExpTask.h>
+#include <tasks/core_0/VescCommsTask.h>
+
 #if SEND_TO_VESC == 0
 #include <tasks/core_0/MockVescTask.h>
 #endif
-#ifdef USING_M5STACK_DISPLAY
+#ifdef USE_M5STACK_DISPLAY
 #include <tasks/core_0/M5StackDisplayTask.h>
 #endif
-#ifdef USING_128x64OLED_TASK
+#ifdef USE_128x64OLED_TASK
 #include <tasks/core_0/I2COledTask.h>
 #endif
-#ifdef USING_IMU
+#ifdef USE_IMU_TASK
 #include <tasks/core_0/IMUTask.h>
 #endif
-#include <tasks/core_0/VescCommsTask.h>
+#ifdef USE_FOOTLIGHT_TASK
 #include <tasks/core_0/FootLightTask.h>
-#include <tasks/core_0/HeadlightTask.h>
-#include <tasks/core_0/I2CPortExpTask.h>
+#endif
 
-#define NUM_TASKS 20
+#define NUM_TASKS 20 // used in taskList
