+++ platformio.ini
@@ -21,10 +21,7 @@ upload_port = COM20
 
 build_flags =
   -D VERSION=4.1
-  -D STANDARDID=\"C53A7D80\"
-  -D M5STACKFIREID=\"8B2DE6B4\"
-  -D TDISPLAYBOARD=\"25286F24\"
-  -D TDISPLAYBOARD_BROWN=\"C9AB6224\"
+
   ; vesc motor stuff
   -D MOTOR_POLE_PAIRS=7
   -D WHEEL_DIAMETER_MM=97
@@ -33,7 +30,12 @@ build_flags =
   -D RPM_AT_MOVING=500
   -D GET_FROM_VESC_INTERVAL=200
   -D NUM_BATT_CELLS=11
+
   -D PRINT_THIS=1
+  -D STANDARDID=\"C53A7D80\"
+  -D M5STACKFIREID=\"8B2DE6B4\"
+  -D TDISPLAYBOARD=\"25286F24\"
+  -D TDISPLAYBOARD_BROWN=\"C9AB6224\"
 
   -D ENABLE_NETWORK_STATS ; used in RF24Network
 
@@ -67,22 +69,22 @@ build_flags =
   -D SEND_TO_VESC=1
   -D MOCK_VESC=0
   -D FOOTLIGHT_PIXEL_PIN=5
-  -D HEADLIGHT_PIN=99 ; change
-  ; -D USING_IMU
+  ; -D USE_IMU_TASK
+  -D USE_FOOTLIGHT_TASK
 
   ; -D RF24_SPI_SPEED=2500000 ; normally use RF24_SPI_SPEED (from NRF config)
 
-[env:T_DISPLAY]
-build_flags = 
-  -D SOFT_SPI_MOSI_PIN=13  # blue
-  -D SOFT_SPI_MISO_PIN=12  # orange
-  -D SOFT_SPI_CLK_PIN=15   # yellow
-  ; -D SPI_CE=26    # green
-  ; -D SPI_CS=27    # white
-  -D SPI_CE=17    # white/violet
-  -D SPI_CS=2    # green
-  -D SOFT_SPI
-  -D FOOTLIGHT_PIXEL_PIN=5 ; not right
+; [env:T_DISPLAY]
+; build_flags = 
+;   -D SOFT_SPI_MOSI_PIN=13  # blue
+;   -D SOFT_SPI_MISO_PIN=12  # orange
+;   -D SOFT_SPI_CLK_PIN=15   # yellow
+;   ; -D SPI_CE=26    # green
+;   ; -D SPI_CS=27    # white
+;   -D SPI_CE=17    # white/violet
+;   -D SPI_CS=2    # green
+;   -D SOFT_SPI
+;   -D FOOTLIGHT_PIXEL_PIN=5 ; not right
 
 
 [env:M5STACKFIRE]
@@ -91,12 +93,10 @@ build_flags =
   -D SPI_CS=13    # white
   -D USING_M5STACK=1
   -D MOCK_VESC=1
-  -D USING_M5STACK_DISPLAY
-  -D USING_128x64OLED_TASK
-  -D USING_IMU
-  ; -D PRINT_DISP_FSM_STATE=1
-  ; -D PRINT_DISP_FSM_TRIGGER=1
-  -D HEADLIGHT_PIN=26
+  -D USE_M5STACK_DISPLAY
+  ; -D USE_128x64OLED_TASK
+  ; -D USE_IMU_TASK
+
   -D FOOTLIGHT_PIXEL_PIN=1
   -D SEND_TO_VESC=0
 
@@ -123,10 +123,11 @@ build_flags =
   # -D PRINT_RX_FROM_CONTROLLER=1
   # -D PRINT_LIGHT_FSM_EVENT_TRIGGER=1
 
-  ; -D FEATURE_FOOTLIGHT
   -D FEATURE_INACTIVITY_FLASH=1
   -D FEATURE_INACTIVITY_TIMEOUT_IN_SECONDS=30
   -D FEATURE_INACTIVITY_FLASH_INTERVAL_IN_SECONDS=10
+
+  ; if using FootlightTask
   -D PRINT_FOOTLIGHT_FSM_STATE=1
   -D PRINT_FOOTLIGHT_FSM_TRIGGER=1
   -D FOOTLIGHT_BRIGHTNESS_MOVING=200
